+++ crates/shared/src/bad_token/mod.rs
@@ -1,6 +1,7 @@
 pub mod cache;
 pub mod instrumented;
 pub mod list_based;
+pub mod token_owner_finder;
 pub mod trace_call;
 
 use anyhow::Result;
